Change from "sticky" elements to "non-sticky".
Instead of using class "sticky" to denote elements that should stick to header when the text is too long, using class "non-sticky" to denote those that shouldn't.  What is actually happening is the "non-sticky" text is hiding behind the header.  This allows the sticky text to remain visible throughout the whole section scroll, not just the h2 tag.